[CIFuzz] Move code for running fuzzers to run_fuzzers.py
TODO: Rename cifuzz.py to build_fuzzers.py
