+++ infra/cifuzz/cifuzz_test.py
@@ -11,10 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests the functionality of the cifuzz module's functions:
-1. Building fuzzers.
-2. Running fuzzers.
-"""
+"""Tests the functionality of the cifuzz module."""
 import os
 import sys
 import tempfile
@@ -28,7 +25,6 @@ sys.path.append(INFRA_DIR)
 OSS_FUZZ_DIR = os.path.dirname(INFRA_DIR)
 
 import cifuzz
-import fuzz_target
 import test_helpers
 
 # NOTE: This integration test relies on
@@ -51,12 +47,6 @@ EXAMPLE_NOCRASH_FUZZER = 'example_nocrash_fuzzer'
 
 # A fuzzer to be built in build_fuzzers integration tests.
 EXAMPLE_BUILD_FUZZER = 'do_stuff_fuzzer'
-
-MEMORY_FUZZER_DIR = os.path.join(TEST_FILES_PATH, 'memory')
-MEMORY_FUZZER = 'curl_fuzzer_memory'
-
-UNDEFINED_FUZZER_DIR = os.path.join(TEST_FILES_PATH, 'undefined')
-UNDEFINED_FUZZER = 'curl_fuzzer_undefined'
 
 # pylint: disable=no-self-use
 
@@ -225,142 +215,6 @@ class BuildFuzzersIntegrationTest(unittest.TestCase):
             'not/a/dir',
             commit_sha='0b95fe1039ed7c38fea1f97078316bfc1030c523',
         ))
-
-
-class RunFuzzerIntegrationTestMixin:  # pylint: disable=too-few-public-methods,invalid-name
-  """Mixin for integration test classes that runbuild_fuzzers on builds of a
-  specific sanitizer."""
-  # These must be defined by children.
-  FUZZER_DIR = None
-  FUZZER = None
-
-  def _test_run_with_sanitizer(self, fuzzer_dir, sanitizer):
-    """Calls run_fuzzers on fuzzer_dir and |sanitizer| and asserts
-    the run succeeded and that no bug was found."""
-    with test_helpers.temp_dir_copy(fuzzer_dir) as fuzzer_dir_copy:
-      run_success, bug_found = cifuzz.run_fuzzers(10,
-                                                  fuzzer_dir_copy,
-                                                  'curl',
-                                                  sanitizer=sanitizer)
-    self.assertTrue(run_success)
-    self.assertFalse(bug_found)
-
-
-class RunMemoryFuzzerIntegrationTest(RunFuzzerIntegrationTestMixin,
-                                     unittest.TestCase):
-  """Integration test for build_fuzzers with an MSAN build."""
-  FUZZER_DIR = MEMORY_FUZZER_DIR
-  FUZZER = MEMORY_FUZZER
-
-  @unittest.skipIf(not os.getenv('INTEGRATION_TESTS'),
-                   'INTEGRATION_TESTS=1 not set')
-  def test_run_with_memory_sanitizer(self):
-    """Tests run_fuzzers with a valid MSAN build."""
-    self._test_run_with_sanitizer(self.FUZZER_DIR, 'memory')
-
-
-class RunUndefinedFuzzerIntegrationTest(RunFuzzerIntegrationTestMixin,
-                                        unittest.TestCase):
-  """Integration test for build_fuzzers with an UBSAN build."""
-  FUZZER_DIR = UNDEFINED_FUZZER_DIR
-  FUZZER = UNDEFINED_FUZZER
-
-  @unittest.skipIf(not os.getenv('INTEGRATION_TESTS'),
-                   'INTEGRATION_TESTS=1 not set')
-  def test_run_with_undefined_sanitizer(self):
-    """Tests run_fuzzers with a valid UBSAN build."""
-    self._test_run_with_sanitizer(self.FUZZER_DIR, 'undefined')
-
-
-class RunAddressFuzzersIntegrationTest(RunFuzzerIntegrationTestMixin,
-                                       unittest.TestCase):
-  """Integration tests for build_fuzzers with an ASAN build."""
-
-  @unittest.skipIf(not os.getenv('INTEGRATION_TESTS'),
-                   'INTEGRATION_TESTS=1 not set')
-  def test_new_bug_found(self):
-    """Tests run_fuzzers with a valid ASAN build."""
-    # Set the first return value to True, then the second to False to
-    # emulate a bug existing in the current PR but not on the downloaded
-    # OSS-Fuzz build.
-    with mock.patch.object(fuzz_target.FuzzTarget,
-                           'is_reproducible',
-                           side_effect=[True, False]):
-      run_success, bug_found = cifuzz.run_fuzzers(10, TEST_FILES_PATH,
-                                                  EXAMPLE_PROJECT)
-      build_dir = os.path.join(TEST_FILES_PATH, 'out', 'oss_fuzz_latest')
-      self.assertTrue(os.path.exists(build_dir))
-      self.assertNotEqual(0, len(os.listdir(build_dir)))
-      self.assertTrue(run_success)
-      self.assertTrue(bug_found)
-
-  @unittest.skipIf(not os.getenv('INTEGRATION_TESTS'),
-                   'INTEGRATION_TESTS=1 not set')
-  def test_old_bug_found(self):
-    """Tests run_fuzzers with a bug found in OSS-Fuzz before."""
-    with mock.patch.object(fuzz_target.FuzzTarget,
-                           'is_reproducible',
-                           side_effect=[True, True]):
-      run_success, bug_found = cifuzz.run_fuzzers(10, TEST_FILES_PATH,
-                                                  EXAMPLE_PROJECT)
-      build_dir = os.path.join(TEST_FILES_PATH, 'out', 'oss_fuzz_latest')
-      self.assertTrue(os.path.exists(build_dir))
-      self.assertNotEqual(0, len(os.listdir(build_dir)))
-      self.assertTrue(run_success)
-      self.assertFalse(bug_found)
-
-  def test_invalid_build(self):
-    """Tests run_fuzzers with an invalid ASAN build."""
-    with tempfile.TemporaryDirectory() as tmp_dir:
-      out_path = os.path.join(tmp_dir, 'out')
-      os.mkdir(out_path)
-      run_success, bug_found = cifuzz.run_fuzzers(10, tmp_dir, EXAMPLE_PROJECT)
-    self.assertFalse(run_success)
-    self.assertFalse(bug_found)
-
-  def test_invalid_fuzz_seconds(self):
-    """Tests run_fuzzers with an invalid fuzz seconds."""
-    with tempfile.TemporaryDirectory() as tmp_dir:
-      out_path = os.path.join(tmp_dir, 'out')
-      os.mkdir(out_path)
-      run_success, bug_found = cifuzz.run_fuzzers(0, tmp_dir, EXAMPLE_PROJECT)
-    self.assertFalse(run_success)
-    self.assertFalse(bug_found)
-
-  def test_invalid_out_dir(self):
-    """Tests run_fuzzers with an invalid out directory."""
-    run_success, bug_found = cifuzz.run_fuzzers(10, 'not/a/valid/path',
-                                                EXAMPLE_PROJECT)
-    self.assertFalse(run_success)
-    self.assertFalse(bug_found)
-
-
-class ParseOutputTest(unittest.TestCase):
-  """Tests parse_fuzzer_output."""
-
-  def test_parse_valid_output(self):
-    """Checks that the parse fuzzer output can correctly parse output."""
-    test_output_path = os.path.join(TEST_FILES_PATH,
-                                    'example_crash_fuzzer_output.txt')
-    test_summary_path = os.path.join(TEST_FILES_PATH, 'bug_summary_example.txt')
-    with tempfile.TemporaryDirectory() as tmp_dir:
-      with open(test_output_path, 'rb') as test_fuzz_output:
-        cifuzz.parse_fuzzer_output(test_fuzz_output.read(), tmp_dir)
-      result_files = ['bug_summary.txt']
-      self.assertCountEqual(os.listdir(tmp_dir), result_files)
-
-      # Compare the bug summaries.
-      with open(os.path.join(tmp_dir, 'bug_summary.txt')) as bug_summary:
-        detected_summary = bug_summary.read()
-      with open(test_summary_path) as bug_summary:
-        real_summary = bug_summary.read()
-      self.assertEqual(detected_summary, real_summary)
-
-  def test_parse_invalid_output(self):
-    """Checks that no files are created when an invalid input was given."""
-    with tempfile.TemporaryDirectory() as tmp_dir:
-      cifuzz.parse_fuzzer_output(b'not a valid output_string', tmp_dir)
-      self.assertEqual(len(os.listdir(tmp_dir)), 0)
 
 
 class CheckFuzzerBuildTest(unittest.TestCase):
